server_authentication/requirements.txt
django
gunicorn
graphene_django
django-heroku
django-graphql-jwt
django-cors-headers
pyjwt==1.7.0
django-silk